+++ src/components/SQForm/useForm.tsx
@@ -4,8 +4,9 @@ import {
   FieldInputProps,
   FieldMetaProps,
   getIn,
-  useField
+  useField,
 } from 'formik';
+import isEqual from 'lodash.isequal';
 import WarningIcon from '@material-ui/icons/NewReleases';
 import VerifiedIcon from '@material-ui/icons/VerifiedUser';
 
@@ -48,15 +49,7 @@ function _handleError(name: string, isRequired: boolean) {
   }
 }
 
-function _getIsFulfilled(hasValue: boolean, isError: boolean) {
-  if (hasValue && !isError) {
-    return true;
-  }
-
-  return false;
-}
-
-function _getHasValue(meta: unknown) {
+function _getHasValue(meta: FieldMetaProps<unknown>) {
   const fieldValue = getIn(meta, 'value');
 
   if (Array.isArray(fieldValue)) {
@@ -78,21 +71,37 @@ export function useForm({
   name,
   isRequired,
   onBlur,
-  onChange
+  onChange,
 }: UseFormParam): UseFormReturn {
   _handleError(name, isRequired);
 
   const [field, meta, helpers] = useField(name);
   const errorMessage = getIn(meta, 'error');
   const isTouched = getIn(meta, 'touched');
+  const isDirty = !isEqual(meta.initialValue, meta.value);
   const hasValue = _getHasValue(meta);
   const isError = !!errorMessage;
-  const isFieldError = (isTouched || hasValue) && isError;
-  const isFieldRequired = isRequired && !hasValue;
-  const isFulfilled = _getIsFulfilled(hasValue, isError);
+
+  const getFieldStatus = () => {
+    if (isRequired && !hasValue && !isDirty && !isTouched) {
+      return 'REQUIRED';
+    }
+    if (isError) {
+      return 'ERROR';
+    }
+    if (hasValue && !isError && isDirty) {
+      return 'USER_FULFILLED';
+    }
+
+    return 'FULFILLED';
+  };
+
+  const isFieldRequired = getFieldStatus() === 'REQUIRED';
+  const isFieldError = getFieldStatus() === 'ERROR';
+  const isFulfilled = getFieldStatus() === 'USER_FULFILLED';
 
   const handleChange = React.useCallback(
-    event => {
+    (event) => {
       field.onChange(event);
       onChange && onChange(event);
     },
@@ -100,7 +109,7 @@ export function useForm({
   );
 
   const handleBlur = React.useCallback(
-    event => {
+    (event) => {
       field.onBlur(event);
       onBlur && onBlur(event);
     },
@@ -142,8 +151,8 @@ export function useForm({
       isError,
       isFieldError,
       isFieldRequired,
-      isFulfilled
+      isFulfilled,
     },
-    fieldHelpers: {handleBlur, handleChange, HelperTextComponent}
+    fieldHelpers: {handleBlur, handleChange, HelperTextComponent},
   };
 }